requirements.txt
bpython==0.12
Django==1.11.1
djangorestframework==3.5.3
djangorestframework-jsonapi==2.1.0
django-nose
inflection==0.3.1
Pygments==2.2.0
pytz==2017.2
simplejson
six==1.10.0
urllib3==1.21.1
wcwidth==0.1.4
